\name{getChrominfo}

\alias{getChrominfo}

\title{
  Retrieves chromosome lenght information from the Ensembl database
}

\description{
  This function uses a perl script to retrieve the chromosome length
  information from an Ensembl database (using the Ensembl perl API).
}

\usage{
getChrominfo(version, ensemblApi, species, user="anonymous", host="ensembldb.ensembl.org", pass="", outfile, ... )
}

\arguments{
  \item{version}{
    Ensembl version from which the object should be created. E.g.: 67
  }
  \item{ensemblApi}{
    The path to the Ensembl perl API (if not already in the PERL5LIB
  environment variable).
  }
  \item{species}{
    Species. Default is \code{human}.
  }
  \item{user}{
    Username for the data base. Default is \code{anonymous}.
  }
  \item{host}{
    Sever where the data base is running.
  }
  \item{pass}{
    Password of the chosen data base user. Default no password is set.
  }
  \item{outfile}{
    The file name of the output file that is generated by the perl
  script. Defaults to: \code{chrominfo-Ensembl-<version>-<species>.txt}
  }
  \item{...}{
    Arguments to be passed to or from methods.
  }
}

\details{
  Making a \code{\linkS4class{TxDb}} or a EnhancedTxDb object
  from ENSEMBL containing annotation information from ENSEMBL.
  The function utilizes a perl script to retrieve the required
  information from ENSEMBL. To make the function work it is essential to
  have the ENSEMBL Perl API installed on your machine and that the
  provided data base connection is correct.
}

\value{
  Returns either a \code{\linkS4class{EnhancedTxDb}} object containing
  the coordinates and annotation of the provided ENSEMBL version or a
  \code{\linkS4class{TxDb}} providing only the genomic
  coordinates of the transcripts, exons, cds and genes.
}

\author{
  D. Bindreither
}

\seealso{
  \link{TxDb}, \link{GRangesList}
}

\examples{
\dontrun{
  ## TranscriptDb-object
  txdb <- createTxDbFromEnsembl(75, annotInfo=FALSE, chr=19)
  txdb

  ## EnhancedTxDb-object
  etxdb <- createTxDbFromEnsembl(75, chr=19)
  etxdb
}
}
